fix(lens): extract a shared pulam-web server test fixture (testKolu.ts)
standUpAgent (parameterized by seed + activity-yield), browserLink, waitFor, and the disposers/afterEach harness (useDisposers) were near-verbatim copies across reserve.test.ts and localKolu.test.ts. Move them into src/server/testKolu.ts and import from both, so a terminalWorkspaceSurface contract change forces one stand-up edit. fleet.test.ts's separate withAgent is left alone.

Agreed by the lowy ⇄ hickey lens debate (finding hickey-4, raised by hickey). Not pushed or merged.
